language2shape/requirements.txt
addict==2.4.0
anyio==3.4.0
argcomplete==1.12.3
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
bpy==2.91a0
bpy-build==2.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
clip @ git+https://github.com/openai/CLIP.git@573315e83f07b53a61ff5098757e8fc885f1703e
cmake==3.22.1
cmake-generators==1.0.9
cycler==0.11.0
Cython==0.29.2
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
distro==1.6.0
dominate==2.6.0
easydict==1.9
entrypoints==0.3
Flask==2.0.2
fonttools==4.28.3
ftfy==6.0.3
future==0.18.2
future-fstrings==1.2.0
gitdb==4.0.9
GitPython==3.1.26
h5py==2.9.0
HTML4Vision==0.4.3
idna==3.3
imageio==2.4.1
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-packaging==0.11.1
jupyter-server==1.12.1
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lucid==0.2.3
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nose==1.3.7
notebook==6.4.6
numpy==1.21.4
open3d==0.13.0
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
plyfile==0.7
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==3.19.1
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
Pygments==2.10.0
PyOpenGL==3.1.5
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
scikit-image==0.19.0
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
svn==1.0.1
tensorboardX==2.4.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
tomlkit==0.7.2
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trimesh==3.9.35
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.6.0
